Transition to NSInteger/NSUInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/projectcenter/trunk@39798 72102866-910b-0410-8b05-ffd578937521

+++ Modules/Editors/ProjectCenter/SyntaxHighlighter.h
@@ -45,7 +45,7 @@
          * italicFont,
          * boldItalicFont;
 
-  unsigned int lastProcessedContextIndex;
+  NSUInteger lastProcessedContextIndex;
 
   NSRange delayedProcessedRange;
   BOOL    didBeginEditing;
